requirements.txt
# GeminiLoop Python Dependencies

# Core AI
google-generativeai>=0.8.0

# Browser automation
playwright>=1.40.0

# Web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0

# Optional: OpenHands integration (future)
# openhands>=0.9.0
PyGithub>=2.1.1
